requirements.txt
fastapi
uvicorn
pydantic>=2
pyyaml
requests
streamlit
